requirements.txt
astroid==2.0.4
certifi==2018.8.13
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-Login==0.4.1
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysql-connector==2.1.6
Pillow==5.2.0
pylint==2.1.1
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
